Completed

+++ hooks/useAuth.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Alert } from 'react-native';
 import { createAction } from '../utils/createAction';
-import { sleep } from '../utils/sleep';
+import { sleep } from '../utils/Sleep';
 import Storage from './../storage/Storage';
 
 export default function useAuth() {
@@ -38,6 +38,16 @@ export default function useAuth() {
                         ...state,
                         isLogin: { ...action.payload },
                     };
+                case 'PAYMENT':
+                    return {
+                        ...state,
+                        payment: { ...action.payload },
+                    };
+                case 'REMOVE_PAYMENT':
+                    return {
+                        ...state,
+                        payment: undefined,
+                    };
                 default:
                     return state;
             }
@@ -50,22 +60,34 @@ export default function useAuth() {
     const auth = React.useMemo(
         () => ({
             login: async (email, password) => {
-                const user = {
-                    email: email,
-                };
-                // await SecureStorage.setItem('user', JSON.stringify(user));
-                await Storage.setItem("user", JSON.stringify(user));
-                await Storage.removeItem('isLogin')
-                dispatch(createAction('REMOVE_DIRECT_LOGIN'));
-                dispatch(createAction('SET_USER', user));
+                if (email === 'ktpm2018@uit.edu.vn' && password === '123') {
+                    const user = {
+                        email: email,
+                    };
+                    await Storage.setItem("user", JSON.stringify(user));
+                    await Storage.removeItem('isLogin')
+                    dispatch(createAction('REMOVE_DIRECT_LOGIN'));
+                    dispatch(createAction('SET_USER', user));
+                    return true;
+                } else {
+                    Alert.alert("Sai Tài Khoản Hoặc Mật Khẩu");
+                    return false;
+                }
+
             },
             logout: async () => {
                 await Storage.removeItem('user')
                 dispatch(createAction('REMOVE_USER'));
             },
             register: async (email, password) => {
-                await sleep(2000);
-                Alert.alert("Đăng ký thành công");
+                await sleep(1000);
+                if (email !== '' && password !== '') {
+                    Alert.alert("Đăng ký thành công");
+                    return true;
+                } else {
+                    Alert.alert("Đăng ký thất bại");
+                    return false;
+                }
             },
             directLogin: async () => {
                 const isLogin = {
@@ -78,11 +100,23 @@ export default function useAuth() {
                 await Storage.removeItem('isLogin')
                 dispatch(createAction('REMOVE_DIRECT_LOGIN'));
             },
+            payment: async () => {
+                const payment = {
+                    payment: 'true',
+                };
+                await Storage.setItem("payment", JSON.stringify(payment));
+                dispatch(createAction('PAYMENT'));
+            },
+            leavePayment: async () => {
+                await Storage.removeItem('payment')
+                await Storage.removeItem('paymentData')
+                dispatch(createAction('REMOVE_PAYMENT'));
+            },
         }),
         [],
     );
     React.useEffect(() => {
-        sleep(2000).then(() => {
+        sleep(1000).then(() => {
             Storage.getItem('user').then(user => {
                 if (user) {
                     dispatch(createAction('SET_USER', JSON.parse(user)));
